backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# CORS
django-cors-headers==4.3.0

# Channels for WebSocket
channels==4.0.0
daphne==4.0.0

# Database
# SQLite is built-in with Python

# Image handling
Pillow==10.1.0

# Gemini API (optional for chatbot)
google-generativeai==0.3.1

# ASGI server
uvicorn==0.24.0

# Environment variables
python-dotenv==1.0.0
